// AngularDemo/src/app/generated/core/api/v1/model/userDto.ts
/**
 * Example OpenAPI Specification
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDto { 
    /**
     * The uuid of a user (Version 7 UUID)
     */
    uuid: string;
    /**
     * The users\' email address
     */
    email: string;
    /**
     * The users\' firstName
     */
    firstName: string;
    /**
     * The users\' lastName
     */
    lastName: string;
    /**
     * The users\' birthdate. full-date notation as defined by RFC 3339, section 5.6, for example, 2017-07-21
     */
    birthdate: string;
}